+++ fuc/api/pyvcf.py
@@ -8,6 +8,15 @@ import gzip
 from copy import deepcopy
 from . import pybed
 
+CONTIGS = [
+    '1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13',
+    '14', '15', '16', '17', '18', '19', '20', '21', '22', 'X', 'Y', 'M',
+    'chr1', 'chr2', 'chr3', 'chr4', 'chr5', 'chr6', 'chr7', 'chr8',
+    'chr9', 'chr10', 'chr11', 'chr12', 'chr13', 'chr14', 'chr15',
+    'chr16', 'chr17', 'chr18', 'chr19', 'chr20', 'chr21', 'chr22',
+    'chrX', 'chrY', 'chrM'
+]
+
 def read_file(fn):
     """Create a VcfFrame from a VCF file (both zipped and unzipped)."""
     meta = []
@@ -439,3 +448,11 @@ class VcfFrame:
             return ann
         s = self.df.apply(func, axis=1)
         return s
+
+    def sort(self):
+        """Return the sorted VcfFrame."""
+        df = self.df.sort_values(by=['#CHROM', 'POS'], ignore_index=True,
+            key=lambda col: [CONTIGS.index(x) if isinstance(x, str)
+                             else x for x in col])
+        vf = self.__class__(deepcopy(self.meta), df)
+        return vf