+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 * :issue:`3`: Add ``sphinx-issues`` extension to Read the Docs.
 * Add the ``VcfFrame.combine`` and ``VcfFrame.subtract`` methods.
 * Rename the ``BedFrame``, ``FastqFrame``, and ``VcfFrame`` modules to ``pybed``, ``pyfq``, and ``pyvcf``, respectively
+* Add the ``VcfFrame.sort`` method.
 
 0.1.4 (2021-04-21)
 ------------------